inlined ddu, duq and het_net implementations no separate representers, decompositions and decide imports
