+++ src/probly/method/__init__.py
@@ -16,6 +16,7 @@ from probly.method.efficient_credal_prediction import efficient_credal_predictio
 from probly.method.ensemble import ensemble
 from probly.method.evidential.classification import evidential_classification
 from probly.method.evidential.regression import evidential_regression
+from probly.method.het_nets import het_nets
 from probly.method.subensemble import subensemble
 
 __all__ = [
@@ -35,5 +36,6 @@ __all__ = [
     "ensemble",
     "evidential_classification",
     "evidential_regression",
+    "het_nets",
     "subensemble",
 ]
