+++ src/probly/method/bayesian.py
@@ -2,6 +2,6 @@
 
 from __future__ import annotations
 
-from probly.transformation.bayesian import BayesianPredictor, bayesian, register
+from probly.transformation.bayesian import BayesianPredictor, bayesian
 
-__all__ = ["BayesianPredictor", "bayesian", "register"]
+__all__ = ["BayesianPredictor", "bayesian"]
