+++ src/probly/method/dropconnect.py
@@ -2,6 +2,6 @@
 
 from __future__ import annotations
 
-from probly.transformation.dropconnect import DropConnectPredictor, dropconnect, dropconnect_traverser, register
+from probly.transformation.dropconnect import DropConnectPredictor, dropconnect
 
-__all__ = ["DropConnectPredictor", "dropconnect", "dropconnect_traverser", "register"]
+__all__ = ["DropConnectPredictor", "dropconnect"]
