+++ src/probly/method/dropout.py
@@ -2,6 +2,6 @@
 
 from __future__ import annotations
 
-from probly.transformation.dropout import DropoutPredictor, dropout, dropout_traverser, register
+from probly.transformation.dropout import DropoutPredictor, dropout
 
-__all__ = ["DropoutPredictor", "dropout", "dropout_traverser", "register"]
+__all__ = ["DropoutPredictor", "dropout"]
