+++ src/probly/method/duq/torch.py
@@ -1,27 +1,54 @@
-"""Torch implementation of the RBF centroid-head transformation."""
+"""Torch implementation of the DUQ transformation."""
 
 from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, ClassVar
 
 import torch
 from torch import nn
 
-from probly.representation.duq.torch import TorchDUQRepresentation
-from probly.transformation.rbf_centroid_head import RBFCentroidHeadPredictor
+from probly.representation._protected_axis.torch import TorchAxisProtected
 from probly.traverse_nn import nn_compose, nn_traverser
 from pytraverse import TRAVERSE_REVERSED, GlobalVariable, State, singledispatch_traverser, traverse_with_state
 
-from ._common import rbf_centroid_head_generator
+from ._common import (
+    DUQPredictor,
+    DUQRepresentation,
+    create_duq_representation,
+    duq_generator,
+    duq_uncertainty,
+)
+
+if TYPE_CHECKING:
+    from torch import Tensor
+
+
+@create_duq_representation.register(torch.Tensor)
+@dataclass(frozen=True, slots=True)
+class TorchDUQRepresentation(DUQRepresentation, TorchAxisProtected):
+    """DUQ representation backed by a torch tensor."""
+
+    kernel_values: Tensor
+    protected_axes: ClassVar[dict[str, int]] = {"kernel_values": 1}
+
+
+@duq_uncertainty.register(torch.Tensor)
+def torch_duq_uncertainty(kernel_values: torch.Tensor) -> torch.Tensor:
+    r"""Per-sample DUQ uncertainty :math:`1 - \max_c K_c(x)` for torch tensors."""
+    return 1.0 - kernel_values.max(dim=-1).values
+
 
 HEAD_MODULE: GlobalVariable[nn.Module | None] = GlobalVariable("RBF_CENTROID_HEAD_MODULE", default=None)
 
 
 @singledispatch_traverser
-def torch_rbf_centroid_head_traverser(obj: nn.Module, state: State) -> tuple[nn.Module, State]:
+def torch_duq_traverser(obj: nn.Module, state: State) -> tuple[nn.Module, State]:
     """Default handler: return module unchanged."""
     return obj, state
 
 
-@torch_rbf_centroid_head_traverser.register
+@torch_duq_traverser.register
 def _(obj: nn.Linear, state: State) -> tuple[nn.Module, State]:
     """Replace the classification head (last Linear, hit first under TRAVERSE_REVERSED) with Identity."""
     if state[HEAD_MODULE] is None:
@@ -136,9 +163,9 @@ class RBFCentroidHead(nn.Module):
         self.centroids_sum.mul_(self.gamma).add_(class_sums, alpha=1.0 - self.gamma)
 
 
-@rbf_centroid_head_generator.register(nn.Module)
-class TorchRBFCentroidHeadPredictor(nn.Module, RBFCentroidHeadPredictor[[torch.Tensor], TorchDUQRepresentation]):
-    """Torch implementation of an RBF centroid-head predictor.
+@duq_generator.register(nn.Module)
+class TorchDUQPredictor(nn.Module, DUQPredictor[[torch.Tensor], TorchDUQRepresentation]):
+    """Torch implementation of an DUQ predictor.
 
     The traversal replaces the last ``nn.Linear`` (the classification head) with
     ``nn.Identity()``, producing a pure feature encoder. A fresh
@@ -174,7 +201,7 @@ class TorchRBFCentroidHeadPredictor(nn.Module, RBFCentroidHeadPredictor[[torch.T
         super().__init__()
         encoder, state = traverse_with_state(
             model,
-            nn_compose(torch_rbf_centroid_head_traverser, nn_traverser=nn_traverser),
+            nn_compose(torch_duq_traverser, nn_traverser=nn_traverser),
             init={TRAVERSE_REVERSED: True, HEAD_MODULE: None},
         )
         head: nn.Linear | None = state[HEAD_MODULE]  # ty: ignore[invalid-assignment]
