+++ src/probly/quantification/__init__.py
@@ -15,7 +15,6 @@ from .decomposition import (
     AleatoricEpistemicTotalDecomposition,
     CachingDecomposition,
     CredalSetEntropyDecomposition,
-    DDUDensityDecomposition,
     Decomposition,
     EpistemicDecomposition,
     SecondOrderEntropyDecomposition,
@@ -40,7 +39,6 @@ __all__ = [
     "AleatoricUncertainty",
     "CachingDecomposition",
     "CredalSetEntropyDecomposition",
-    "DDUDensityDecomposition",
     "Decomposition",
     "EpistemicDecomposition",
     "EpistemicUncertainty",
