--- src/probly/representation/het_nets/__init__.py
@@ -1,9 +1,0 @@
-"""HetNets representation."""
-
-from __future__ import annotations
-
-from ._common import HetNetsRepresentation
-
-__all__ = [
-    "HetNetsRepresentation",
-]
