--- tests/probly/quantification/decomposition/ddu/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for DDU uncertainty decompositions."""
